Assistant checkpoint: Enhanced portfolio with improved styling and features
Assistant generated file changes:
- templates/portfolio.html: Enhanced portfolio page with improved styling and additional features

---

User prompt:

in the footer include a statement to inform the users/employers that am also available to work remotely or part time incase they consider to hire me. include a go back button so that after viewing my portfolio users are able to return to their previous page. also make the whatsapp icon green. let the scheme colour have some dark and light places in the whole portfolio page .in the technical skills part on the portfolio page let each of the skill have a background image choose a relevant image from the internet  and place on each one of them. throughout the portfolio page use the font the colour and the style you want to style the texts to be more professional. also style the header of the portfolio page position it professionally so it is appealing to users/employers. on your own position style the whole portfolio page  to look the best the best way you can. include any other relevant detail i might have forgotten that is relevant to the portfolio.