Assistant checkpoint: Update portfolio with enhanced styling and content
Assistant generated file changes:
- templates/portfolio.html: Add hover effects to social icons

---

User prompt:

in the featured projects include my youtube channel, https://www.youtube.com/channel/UC6d9bq2T8A72bl6zrZvT_wQ, also in the featured projects include our logo and for Absolute Beginners motto.in the header make the background image a bit more visible so as it is attractive and draws attention to the user/employer also on your own design this header to stop being boring remember it is the first thing the user/employer sees when they click to view our portfolio you are free to use relevant ways. make the icons in the get in touch part react when the mouse hoover on them. put this url,https://github.com/Lilian-Powers/CodeCraft- ,  such that when the user click view us on github they are redirected there.