Assistant checkpoint: Update navigation layout and access control
Assistant generated file changes:
- app.py: Update home route to allow access without login
- templates/base.html: Move navigation buttons and update nav, Add floating navigation buttons
- templates/tutorials.html: Update tutorials heading

---

User prompt:

make it so when the users get on our site let them land on the home page first and then they are redirected to the login/register when they click to access the links if they have not yet login/registered. Let the back/next arrow be on the bottom right of each page, remove them from the nav bar since there they appear hidden and we really want them where the user can see them all the time so that they can use them incase they want to go back to the previous page or the next one. Also make it so when a user logs in and they are not yet registered let them be redirected to the register page. make this heading shorter and interesting since it is repeating what is written in the following text use the font and the style you want to make it look good to the users who click to start coding.