Assistant checkpoint: Add interactive food favorites game with modal
Assistant generated file changes:
- templates/tutorials.html: Add favorite food game with modal, Add food game styles, Add food game JavaScript

---

User prompt:

create a super hero name generator game using this code, first_name = food = input("What is your favorite food? ")
print(f"Yum! I love {food} too!"), for our learners/users. make so it is like an attractive and appealing game to the users/learners to encourage them to continue with the other projects, displays nice compliments when the favourite food is displayed to make the user feel good/nice. make it so it is in a modal so that it pops up above the other elements and is centrallay positioned. use the font, colour or style that you think is the best to make the game appealing and attractive to the users.