Assistant checkpoint: Add Funny Story Maker game with modal interface
Assistant generated file changes:
- templates/tutorials.html: Add Funny Story Maker game, Add Story Maker styles, Add Story Maker JavaScript functions

---

User prompt:

create a funny story maker game using this code,name = input("Enter a name: ")
place = input("Enter a place: ")
activity = input("Enter an activity: ")

print(f"Once upon a time, {name} went to {place} to {activity}. It was so much fun!"), for our learners/users. make so it is like an attractive and appealing game to the users/learners to encourage them to continue with the other projects, displays nice compliments when the story is displayed to make the user feel good/nice about themsalves and the story they have just made. make it so it is in a modal so that it pops up above the other elements and is centrallay positioned. use the font, colour or style that you think is the best to make the game appealing and attractive to the users.