Assistant checkpoint: Fix modal positioning and click outside behavior
Assistant generated file changes:
- templates/tutorials.html: Fix modal positioning and click outside behavior, Fix modal overlay and click outside behavior

---

User prompt:

we have at this for a long time you keep telling me you have fixed yet the issues are still there for example we said we fix the issue so that when i click outside all the open modals in the tutorial page can close automatically you haven't done that i am trying to click outside the open modal is not closing. another is that we have been over this issue of  the modals opening in the wrong positions i just want any modal that i click to open in the center of the page when i click and to stop going all the way uto the top of the page but to open directly above the other projects. for the last 2 hours we have been going back and forth on these same issues why are they so difficult for you to fix. if there is code to remove or add or change or any other approach to make it  so that the modals can display properly at  the center without being hidden.