Assistant checkpoint: Add explanatory note about game limitations
Assistant generated file changes:
- templates/tutorials.html: Add note about compliment generator limitations

---

User prompt:

directly below the practice projects items, add italisized text informing the  users that some of the above compliments from the games may not make sence because they are fixed for example some might be even gramatically wrong because the code we are using is so simple that it does not offer the computer more options to choose from so that it can give use a perfect answer. acknowledge that this is a great improvement from the games we started with and as we move on to part 2 of the tutorials and learn more stuff in python this will continue to be more perfect and interesting.style and position the text professionally use the font color ans style that you think is best.