Assistant checkpoint
Assistant generated file changes:
- templates/portfolio.html: Update carousel images and captions to match correctly

---

User prompt:

i know where the issue is, go back to the static folder and rename the images for that  you are using for the carousel you did not name them well and that is why you are having miss match of the caption and the images. also make sure that all the images that i told you to put in the carousel are there are are being displayed correctly.