Assistant checkpoint: Update navigation, reorganize content, and improve UI
Assistant generated file changes:
- templates/portfolio.html: Reorder carousel images for better visual appeal
- templates/base.html: Update navigation menu items
- templates/tutorials_part2.html: Update calculator page title
- templates/home.html: Update quick links section
- templates/tutorials.html: Update tutorials page with sections, Add section styles and script

---

User prompt:

organize the images in the carousel with the most attractive to start so that they can attract and maintain the users attention and be appealing to them. lets change things a bit, rename the python tutorial part 2 page to a nice short name about the calculator so that that page can be for the calculator only for now, have a link for it in the quck links and in the nav bar. then let all the tutorials be in one page the one named tutorials part1 give it another nice short name but make it so it has sections like all the content in the tutorials part1 page right now will be in section 1 and the tutorials for section 2 will be coming soon organise style use fonts and colours that you think are the best to style and position elements in this page professionally so that it is easier on the user/learners, have a link for it in the quck links and in the nav bar.