Assistant checkpoint: Update tutorials page layout and styling
Assistant generated file changes:
- templates/tutorials.html: Remove duplicate note container and update Part 2 section, Update Video Tutorials Part 2 section, Update tutorials section 2 styles

---

User prompt:

go through the tutorials.html page code remove dublicate code that we are not using for the website. also where did this go, in the tutorials page directly above the footer let us have Video Tutorials Part 2 use the style font colour that you like to position professionally. inform the user/learner that this part they continue to practice what they have learnt and also learn introduction to conditional statements if elif and else using these short YouTube videos, https://youtu.be/hginwuW4Lw0, and, https://youtu.be/XivtThGig-Q, position and stle them professionally the best way you can make it so the videos can play within the website give the nice and appealing headers, because i want ti to be displayed directly above the footer in the tutorials.html page but leve some space between the footer an it. i still cannot see it display on the website. where is the problem